+++ watson/utils.py
@@ -32,10 +32,13 @@ except NameError:
 def confirm_project(project, watson_projects):
     """
     Ask user to confirm creation of a new project
+    'project' must be a string
+    'watson_projects' must be an interable.
     Returns True on accept and raises click.exceptions.Abort on reject
     """
     if project not in watson_projects:
-        msg = "Project '%s' does not exist yet. Create it?" % project
+        msg = ("Project '%s' does not exist yet. Create it?"
+               % style('project', project))
         click.confirm(msg, abort=True)
     return True
 
@@ -43,11 +46,12 @@ def confirm_project(project, watson_projects):
 def confirm_tags(tags, watson_tags):
     """
     Ask user to confirm creation of new tags (each separately)
+    Both 'tags' and 'watson_tags" must be iterables.
     Returns True if all accepted and raises click.exceptions.Abort on reject
     """
     for tag in tags:
         if tag not in watson_tags:
-            msg = "Tag '%s' does not exist yet. Create it?" % tag
+            msg = "Tag '%s' does not exist yet. Create it?" % style('tag', tag)
             click.confirm(msg, abort=True)
     return True